giovanni ayuda

+++ VentanaPrincipal/GestionUsuarios/frmGestionUsuarios.cs
@@ -79,9 +79,7 @@ namespace GGNoTeam_V5.VentanaPrincipal
         }
 
         private void btnUsuarios_Click(object sender, EventArgs e)
-        {
-            dgvPersonas.Rows.Clear();
-            Global.pintarDGV(ref dgvPersonas, btnListarUsuarios.BackColor);
+        {                        
             LoginWS.persona[] usuarios = _daoListar.listarUsuarios();
             if (usuarios != null)
             {
@@ -90,14 +88,13 @@ namespace GGNoTeam_V5.VentanaPrincipal
             }
             else
             {
-
+                dgvPersonas.DataSource = null;
             }
+            Global.pintarDGV(ref dgvPersonas, btnListarUsuarios.BackColor);
         }
 
         private void btnAdministradores_Click(object sender, EventArgs e)
-        {
-            Global.pintarDGV(ref dgvPersonas, btnListarAdministradores.BackColor);
-            dgvPersonas.Rows.Clear();
+        {                   
             LoginWS.persona[] administradores = _daoListar.listarAdministradores();
             if (administradores != null)
             {
@@ -106,22 +103,18 @@ namespace GGNoTeam_V5.VentanaPrincipal
             }
             else
             {
-
+                dgvPersonas.DataSource = null;
             }
+            Global.pintarDGV(ref dgvPersonas, btnListarAdministradores.BackColor);
         }
 
         private void colocarEnDGV(LoginWS.persona[] lista)
         {
-            for (int i = 0; i < lista.Length; i++)
-            {
-                dgvPersonas.Rows.Add(lista[i].idPersona, lista[i].codigo, lista[i].nombre, lista[i].apellidos, lista[i].correo, lista[i].contraseña, lista[i].validacion, lista[i].rol);
-            }
+            dgvPersonas.DataSource = lista;
         }
 
         private void btnPersona_Click(object sender, EventArgs e)
-        {
-            Global.pintarDGV(ref dgvPersonas, btnListarPersonas.BackColor);
-            dgvPersonas.Rows.Clear();
+        {                     
             LoginWS.persona[] noUsuarios = _daoListar.listarNoUsuarios();
             if (noUsuarios != null)
             {
@@ -130,26 +123,50 @@ namespace GGNoTeam_V5.VentanaPrincipal
             }
             else
             {
-
+                dgvPersonas.DataSource = null;
             }
-
-                    
+            Global.pintarDGV(ref dgvPersonas, btnListarPersonas.BackColor);
         }
 
         private void btnEliminarUsuario_Click(object sender, EventArgs e)
         {
-            _daoListar.eliminarPersona(lista[dgvPersonas.CurrentRow.Index].idPersona, lista[dgvPersonas.CurrentRow.Index].itinerario.idItineraio);
-            //MessageBox.Show("¿Está seguro que desea eliminar a la persona?", MessageBoxButtons.OKCancel);
+            int resultadoEliminacion = -1;
+            if(dgvPersonas.CurrentRow != null)
+            {
+                
+                var result = MessageBox.Show("¿Está seguro que desea eliminar a la persona?","Eliminar persona", MessageBoxButtons.OKCancel);
+                if(result == DialogResult.OK)
+                {
+                    resultadoEliminacion = _daoListar.eliminarPersona(lista[dgvPersonas.CurrentRow.Index].idPersona, lista[dgvPersonas.CurrentRow.Index].itinerario.idItineraio);
+                    if(resultadoEliminacion == 1)
+                    {
+                        MessageBox.Show("Se ha eliminado correctamente al usuario");
+                        this.btnPersona_Click(sender, e);
+                    } else
+                    {
+                        MessageBox.Show("Hubo un error al eliminar al usuario");
+                    }
+                } else if (result == DialogResult.Cancel)
+                {
+                    return;
+                }
+            }
+
+
         }
 
 
 
         private void btnActualizarUsuario_Click(object sender, EventArgs e)
         {
-            dgvColorActivo = dgvPersonas.GridColor;
-            desactivarBotones();
-            frmEditarUsuario editarUsuarios = new frmEditarUsuario(lista[dgvPersonas.CurrentRow.Index], btnActualizarUsuario.BackColor, this);
-            editarUsuarios.ShowDialog();
+            if (dgvPersonas.CurrentRow != null)
+            {
+                dgvColorActivo = dgvPersonas.GridColor;
+                desactivarBotones();
+
+                frmEditarUsuario editarUsuarios = new frmEditarUsuario(lista[dgvPersonas.CurrentRow.Index], btnActualizarUsuario.BackColor, this);
+                editarUsuarios.ShowDialog();
+            }
         }
 
         private void btnAgregarUsuario_Click(object sender, EventArgs e)
@@ -174,8 +191,7 @@ namespace GGNoTeam_V5.VentanaPrincipal
         {
             if (txtBoxBusqueda.Texts != "Ingrese el nombre o el código del usuario:")
             {
-                Global.pintarDGV(ref dgvPersonas, Color.SlateBlue);
-                dgvPersonas.Rows.Clear();
+                Global.pintarDGV(ref dgvPersonas, Color.SlateBlue);                
                 lista = _daoListar.listarPorCodNom(txtBoxBusqueda.Texts);
                 if (lista != null)
                 {
@@ -206,6 +222,25 @@ namespace GGNoTeam_V5.VentanaPrincipal
         private void txtBoxBusqueda_Leave(object sender, EventArgs e)
         {
             //txtBoxBusqueda.Texts = "Ingrese el nombre o el código del usuario:";
+        }
+
+        private void label1_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void dgvPersonas_CellFormatting(object sender, DataGridViewCellFormattingEventArgs e)
+        {
+            LoginWS.persona user = (LoginWS.persona)dgvPersonas.Rows[e.RowIndex].DataBoundItem;
+
+            dgvPersonas.Rows[e.RowIndex].Cells[0].Value = user.idPersona;
+            dgvPersonas.Rows[e.RowIndex].Cells[1].Value = user.codigo;
+            dgvPersonas.Rows[e.RowIndex].Cells[2].Value = user.nombre;
+            dgvPersonas.Rows[e.RowIndex].Cells[3].Value = user.apellidos;
+            dgvPersonas.Rows[e.RowIndex].Cells[4].Value = user.correo;
+            dgvPersonas.Rows[e.RowIndex].Cells[5].Value = user.contraseña;
+            dgvPersonas.Rows[e.RowIndex].Cells[6].Value = user.validacion;
+            dgvPersonas.Rows[e.RowIndex].Cells[7].Value = user.rol;
         }
     }
 }
